+++ lib/CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2023-06-27
+### Added
+- Launch flag for initial open lightbox
+- Prevent close flag for close lightbox
+- Target attribute for close button
+
+### Changed
+- Automatic close after events
+- Optimized build (unique CSS file)
+- More descriptive event names
+
 ## [1.2.0] - 2022-11-23
 ### Added
 - Playground and styles